fix: preserve presale flow and rename Treasure Hunt route
