+++ dapp/__tests__/lib/wallet-signed-session.test.ts
@@ -4,14 +4,14 @@ import {
   type WalletSessionPayload,
 } from '@/lib/wallet-auth';
 
-const buyerAddress = '0x1111111111111111111111111111111111111111';
+const walletAddress = '0x1111111111111111111111111111111111111111';
 const otherAddress = '0x2222222222222222222222222222222222222222';
 
 function session(overrides: Partial<WalletSessionPayload> = {}): WalletSessionPayload {
   return {
-    userId: 'buyer-user',
-    walletAddress: buyerAddress,
-    signedWalletAddress: buyerAddress,
+    userId: 'wallet-user',
+    walletAddress,
+    signedWalletAddress: walletAddress,
     walletType: 'evm',
     issuedAt: '2026-07-16T00:00:00.000Z',
     expiresAt: '2026-08-16T00:00:00.000Z',
@@ -19,9 +19,9 @@ function session(overrides: Partial<WalletSessionPayload> = {}): WalletSessionPa
   };
 }
 
-describe('autorizacion EVM de referrals', () => {
+describe('sesion EVM firmada', () => {
   it('solo acepta una direccion EVM completa', () => {
-    expect(isValidEvmWalletAddress(buyerAddress)).toBe(true);
+    expect(isValidEvmWalletAddress(walletAddress)).toBe(true);
     expect(isValidEvmWalletAddress('0xbuyer')).toBe(false);
     expect(isValidEvmWalletAddress('0x0000000000000000000000000000000000000000')).toBe(false);
     expect(isValidEvmWalletAddress('TJRabPrwbZy45sbavfcjinPJC18kjpRTv8')).toBe(false);
@@ -29,19 +29,19 @@ describe('autorizacion EVM de referrals', () => {
 
   it('compara exclusivamente con signedWalletAddress', () => {
     expect(evmWalletSessionMatchesSignedAddress(session({
-      walletAddress: buyerAddress,
+      walletAddress,
       signedWalletAddress: otherAddress,
-    }), buyerAddress)).toBe(false);
+    }), walletAddress)).toBe(false);
   });
 
   it('rechaza una sesion TRON aunque el campo firmado contenga la direccion solicitada', () => {
     expect(evmWalletSessionMatchesSignedAddress(session({
-      signedWalletAddress: buyerAddress,
+      signedWalletAddress: walletAddress,
       walletType: 'tron',
-    }), buyerAddress)).toBe(false);
+    }), walletAddress)).toBe(false);
   });
 
   it('acepta la wallet EVM que firmo la sesion', () => {
-    expect(evmWalletSessionMatchesSignedAddress(session(), buyerAddress)).toBe(true);
+    expect(evmWalletSessionMatchesSignedAddress(session(), walletAddress)).toBe(true);
   });
 });
