+++ dapp/src/app/api/chat/init/route.ts
@@ -8,8 +8,8 @@ export async function POST(request: NextRequest) {
     const rooms = [
       {
         gameId: 'sybil-slayer',
-        name: 'Sybil Slayer Chat',
-        description: 'Chat room for Sybil Slayer game players',
+        name: 'Treasure Hunt Chat',
+        description: 'Chat room for Treasure Hunt players',
       },
       {
         gameId: 'hyppie-road',
