+++ games/sybil-slayer/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "name": "Sybil Slayer - Treasure Hunt",
-  "short_name": "Sybil Slayer",
+  "name": "Treasure Hunt",
+  "short_name": "Treasure Hunt",
   "description": "¡Esquiva duendes y recolecta puntos!",
   "start_url": "/",
   "display": "fullscreen",
@@ -31,4 +31,3 @@
     }
   ]
 }
-
